+++ flang/lib/Parser/unparse.cpp
@@ -2758,7 +2758,7 @@ public:
   }
   void Unparse(const OpenMPInvalidDirective &x) {
     BeginOpenMP();
-    Word("!$OMP ");
+    Word(x.isExtensionSentinel ? "!$OMPX " : "!$OMP ");
     Put(parser::ToUpperCaseLetters(x.source.ToString()));
     Put("\n");
     EndOpenMP();